adapter/requirements.txt
mysql.connector-python
influxdb
get_docker_secret